Multi-line svn commit, see body.
Pass the whole document to the transform, not just the single element.  This means that stylesheets written for use with document-level tools like command-line Xalan will work without change in this environment.
This is now safe to do because a DomElementItem's wrapped Element is now guaranteed to be the document element of its owning Document.