Multi-line svn commit, see body.
XSLTransformationStage now handles only the whole document -> whole document case, which is the most usual one.  This allows it to do a better job (for example, allowing comments and processing elements to be generated outside the document element) and simplifies its implementation to some extent as well.
The original multi-output case is renamed as MultiOutputXSLTransformationStage.
Both implementations have unit tests of their own.