MDA-113: review code with -Xlint:unchecked
Fixed all visible problems in the main code.  I have *not* done the same for the tests as yet (and probably never will) although I did take a look at a few.
One minor API change came out of this, where a parameter which was a "..." of a generic type has been changed to a List<Type>.  This is almost certainly not visible to users, unless they have developed alternate implementations of CollectionMergeStrategy (which seems unlikely).