MDA-216 - remove CRs from signatures made under Java 11
Relaxed the canary in this test so as to handle the fact that in Oracle Java 8u231 (but not in equivalent OpenJDKs, etc.) the same change has been back-ported.

Now, all we can really check for is that IF there's a difference when we turn off the option, some CRs have appeared.