MDA-242 - Review for thread safety
https://issues.shibboleth.net/jira/browse/MDA-242

Additional issues highlighted by SpotBugs.

Note that making the doInitialize and doDestroy methods synchronized isn't actually ever required, as they are executed as a chain by a public method in a parent class which _is_ synchronized. SpotBugs thinks it is because it is unaware of the context in which these are called.

However, this change doesn't cost us much (double synchronization is supposedly inexpensive) and does allow SpotBugs to operate on the whole codebase without fiddly filtering of specific issues against specific method names.