MDA-275 - Default stage configurations should not create daemon threads
Change API of three stages from an ExecutorService to an Executor.

This simplification allows a switch to a DirectExecutor by default,
which just executes tasks on the calling thread. This means we don't
allocate any thread pools by default, leaving that to the caller if it
is to be done. It also means we're not creating any non-daemon threads
by default.

https://shibboleth.atlassian.net/browse/MDA-275